fix windows build error
