+++ zlgcan/src/can/channel/mod.rs
@@ -10,11 +10,60 @@ mod windows;
 #[cfg(target_os = "windows")]
 pub(crate) use windows::*;
 
-use std::ffi::c_uint;
+use std::collections::HashMap;
+use std::ffi::{c_uchar, c_uint, c_ushort};
 use rs_can::{CanError, ChannelConfig};
-use crate::can::{common::BitrateCtx, ZCanFilterType};
-use crate::{ACC_CODE, ACC_MASK, BRP, CHANNEL_MODE, FILTER_TYPE};
+use crate::{can::{common::BitrateCtx, ZCanFilterType, constant::{BRP, SJW, SMP, TSEG1, TSEG2}}, constants};
 use crate::device::ZCanDeviceType;
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub(crate) struct ZCanFdChlCfgSet {
+    tseg1: c_uchar,
+    tseg2: c_uchar,
+    sjw: c_uchar,
+    smp: c_uchar,
+    brp: c_ushort,
+}
+
+impl TryFrom<&HashMap<String, u32>> for ZCanFdChlCfgSet {
+    type Error = CanError;
+    fn try_from(value: &HashMap<String, u32>) -> Result<Self, Self::Error> {
+        let &tseg1 = value.get(TSEG1)
+            .ok_or(CanError::OtherError(format!("`{}` is not configured in file!", TSEG1)))?;
+        let &tseg2 = value.get(TSEG2)
+            .ok_or(CanError::OtherError(format!("ZLGCAN - `{}` is not configured in file!", TSEG2)))?;
+        let &sjw = value.get(SJW)
+            .ok_or(CanError::OtherError(format!("ZLGCAN - `{}` is not configured in file!", SJW)))?;
+        let &smp = value.get(SMP)
+            .ok_or(CanError::OtherError(format!("ZLGCAN - `{}` is not configured in file!", SMP)))?;
+        let &brp = value.get(BRP)
+            .ok_or(CanError::OtherError(format!("ZLGCAN - `{}` is not configured in file!", BRP)))?;
+
+        Ok(Self::new(tseg1, tseg2, sjw, smp, brp))
+    }
+}
+
+impl ZCanFdChlCfgSet {
+    #[inline(always)]
+    pub fn new(tseg1: u32, tseg2: u32, sjw: u32, smp: u32, brp: u32) -> Self {
+        Self {
+            tseg1: tseg1 as u8,
+            tseg2: tseg2 as u8,
+            sjw: sjw as u8,
+            smp: smp as u8,
+            brp: brp as u16,
+        }
+    }
+    /// Only used for USBCANFD-800U
+    #[inline(always)]
+    pub fn get_timing(&self) -> u32 {
+        (self.brp as u32) << 22
+            | (self.sjw as u32 & 0x7f) << 15
+            | (self.tseg2 as u32 & 0x7f) << 8
+            | (self.tseg1 as u32)
+    }
+}
 
 #[repr(C)]
 pub(crate) struct ZCanChlCfg {
@@ -36,15 +85,15 @@ impl ZCanChlCfg {
                 can_type: can_type as u32,
                 cfg: ZCanChlCfgUnion {
                     canfd: common::ZCanFdChlCfgInner::new(
-                        cfg.get_other::<u8>(CHANNEL_MODE)?
+                        cfg.get_other::<u8>(constants::CHANNEL_MODE)?
                             .unwrap_or(ZCanChlMode::Normal as u8),
                         aset.get_timing(),  // timing0 and timing1 ignored expect USBCANFD_800U
                         dset.get_timing(),
-                        cfg.get_other::<u8>(FILTER_TYPE)?
+                        cfg.get_other::<u8>(constants::FILTER_TYPE)?
                             .unwrap_or(ZCanFilterType::default() as u8),
-                        cfg.get_other::<u32>(ACC_CODE)?,
-                        cfg.get_other::<u32>(ACC_MASK)?,
-                        cfg.get_other::<u32>(BRP)?,
+                        cfg.get_other::<u32>(constants::ACC_CODE)?,
+                        cfg.get_other::<u32>(constants::ACC_MASK)?,
+                        cfg.get_other::<u32>(constants::BRP)?,
                     )?
                 }
             })
